updates easy ads repository structure and made example worked
