+++ example/lib/screens/country_list_screen.dart
@@ -1,8 +1,6 @@
 import 'package:ads/screens/country_detail_screen.dart';
 import 'package:ads/models/country.dart';
-import 'package:ads/utils/easy_ads/easy_ad_base.dart';
-import 'package:ads/utils/easy_ads/easy_admob/easy_admob_banner_ad.dart';
-import 'package:ads/utils/easy_ads/easy_ads.dart';
+import 'package:easy_ads/easy_ads.dart';
 import 'package:flutter/material.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 
@@ -14,7 +12,7 @@ class CountryListScreen extends StatefulWidget {
 }
 
 class _CountryListScreenState extends State<CountryListScreen> {
-  late final EasyAdBase _bannerAd;
+  late final EasyAdmobBannerAd _bannerAd;
   @override
   void initState() {
     super.initState();
