+++ lib/src/easy_ad_base.dart
@@ -1,5 +1,5 @@
-import 'package:ads/utils/enums/ad_network.dart';
-import 'package:ads/utils/enums/ad_unit_type.dart';
+import 'package:easy_ads/src/enums/ad_network.dart';
+import 'package:easy_ads/src/enums/ad_unit_type.dart';
 
 abstract class EasyAdBase {
   final String adUnitId;
@@ -11,8 +11,6 @@ abstract class EasyAdBase {
   AdUnitType get adUnitType;
   bool get isAdLoaded;
 
-  /// This will be called for initialization when we have to wait for the initialization
-  Future<void> init();
   void dispose();
 
   /// This will load ad, It will only load the ad if isAdLoaded is false
