+++ pubspec.yaml
@@ -11,6 +11,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  google_mobile_ads: ^0.13.4
 
 dev_dependencies:
   flutter_test:
